feat(usecase): validate warehouse id before delete

+++ src/main/kotlin/domain/usecase/crud/warehouse/DeleteWarehouseUseCase.kt
@@ -1,11 +1,18 @@
 package domain.usecase.crud.warehouse
 
 import domain.repository.WarehouseRepository
+import domain.validator.ValidationResult
+import domain.validator.WarehouseValidator
 
 class DeleteWarehouseUseCase(
     private val warehouseRepository: WarehouseRepository
 ) {
-    suspend operator fun invoke(id:String):Boolean {
-    return   warehouseRepository.deleteWarehouse(id)
+    suspend operator fun invoke(id: String): Boolean {
+        val validationResult = WarehouseValidator.validateId(id)
+        if (validationResult is ValidationResult.Failure) {
+            return false
+        }
+
+        return warehouseRepository.deleteWarehouse(id)
     }
 }
